Fixed compiler warnings and test failures found by buildbot Fixed ccfilter to detect errors where the column is included in the error message

+++ mysys/mf_iocache.c
@@ -517,7 +517,7 @@ int _my_b_write(IO_CACHE *info, const uchar *Buffer, size_t Count)
   {
     my_off_t old_pos_in_file= info->pos_in_file;
     res= info->write_function(info, Buffer, Count);
-    Count-= info->pos_in_file - old_pos_in_file;
+    Count-= (size_t) (info->pos_in_file - old_pos_in_file);
     Buffer+= info->pos_in_file - old_pos_in_file;
   }
   else
@@ -1226,7 +1226,7 @@ static int _my_b_cache_read_r(IO_CACHE *cache, uchar *Buffer, size_t Count)
 static void copy_to_read_buffer(IO_CACHE *write_cache,
                                 const uchar *write_buffer, my_off_t pos_in_file)
 {
-  size_t write_length= write_cache->pos_in_file - pos_in_file;
+  size_t write_length= (size_t) (write_cache->pos_in_file - pos_in_file);
   IO_CACHE_SHARE *cshare= write_cache->share;
 
   DBUG_ASSERT(cshare->source_cache == write_cache);
